finish semantic check

+++ src/AST/expression/NumberExp.java
@@ -1,11 +1,9 @@
 package src.AST.expression;
 
-import src.AST.ASTVisitor;
-
 public class NumberExp extends Expression {
-    public int value;
+    public long value;
 
     public NumberExp(String number) {
-        value = Integer.parseInt(number);
+        value = Long.parseLong(number);
     }
 }
